Location code corrected

+++ app/src/main/java/technians/com/vucabsdriver/Model/PendingRequest/PendingRequestData.java
@@ -6,31 +6,31 @@ import java.util.ArrayList;
 
 public class PendingRequestData {
     @SerializedName("total_earn")
-    private int total_earn;
+    private float total_earn;
     @SerializedName("total_booking")
-    private int total_booking;
+    private float total_booking;
     @SerializedName("booking_data")
     private ArrayList<BookingData> bookingData;
 
-    public PendingRequestData(int total_earn, int total_booking, ArrayList<BookingData> bookingData) {
+    public PendingRequestData(float total_earn, float total_booking, ArrayList<BookingData> bookingData) {
         this.total_earn = total_earn;
         this.total_booking = total_booking;
         this.bookingData = bookingData;
     }
 
-    public int getTotal_earn() {
+    public float getTotal_earn() {
         return total_earn;
     }
 
-    public void setTotal_earn(int total_earn) {
+    public void setTotal_earn(float total_earn) {
         this.total_earn = total_earn;
     }
 
-    public int getTotal_booking() {
+    public float getTotal_booking() {
         return total_booking;
     }
 
-    public void setTotal_booking(int total_booking) {
+    public void setTotal_booking(float total_booking) {
         this.total_booking = total_booking;
     }
